requirements.txt
altair
seaborn
SLPP
SQLAlchemy
streamlit
pygsheets
pandas
xlsxwriter
beautifulsoup4
requests
streamlit-flow-component

# for transfering the items database
mysql-to-sqlite3